requirements.txt
langchain-community
langchain-openai
langgraph
tavily-python
requests
beautifulsoup4
python-dotenv